TestAccount
I made a different implementation of test. I added a menu of choices for
client. I'm having some problems can you look at it and let's work on
this later if you want to use it. I think it's better because the client
has the final choice what he wants to do. ttyll